+++ src/App.js
@@ -1,11 +1,13 @@
 import { getNavbar } from "./assets/Navbar/getNavbar.js";
 import { getSearcher } from "./assets/Searcher/getSearcher.js";
 import { getProducts } from "./assets/Products/getProducts.js";
+import { getButtons } from "./assets/Navbar/getButtons.js";
 
 export const init = async () => {
     
   getSearcher();
-  await getNavbar();
+  getNavbar();
+  await getButtons();
   await getProducts();
 
 }
